requirements.txt
requests
beautifulsoup4
celery
lxml
psycopg2-binary
datetime